Update: ThirdParty Dependencies
:construction: :tophat: :construction:

Added third party git repos as submodules for this game engine. Committing now as they appear to be in the correct location, in the correct state and are pointing to the correct branches.

Any updates to the premake files will need to be make in their respective repositories - hopefully I've already set them up to use the correct filepaths and includes but I will have to wait and see whilst I get things setup.