Update: Premakes + Shared Project
Added new shared project which should build all of the libraries/third party projects before linking everything in the engine project.

This is so that all of the files/folders are available when building the engine and explicitly linking .lib files (hopefully to fix the linux linking errors).

Need to experiment with the make command though, need to try and build specific projects using make, not sure how that's done, hopefully specifying "project" is enough.