Fix: Release/Distribution Executable Crash
ASSERTs will be compiled out in anything that isn't DEBUG at the moment. I had bgfx initializing in an assert. So when CI came to test the release executable, that bgfx initialize was already compiled out. Should be handling scenarios like this by storing the return value and then asserting that. :no_entry_sign: