Update: ImGui Fonts
Reverting to the default font for now. Can maybe try specific fonts later, but this is causing the Linux/Mac exes to crash because it's the wrong filename. Not sure if this has to be the absolute filepath or just the font filename (I can test this at some point), but honestly, not that important right now.