Update: Application Layers + Cleanup
Crash could be due to the order of deletion perhaps? In the test repo the ImGui layer is removed first and then the window destroys. I'm not sure on the order of destruction here with these pointers and variables, but I should be popping the ImGui layer off before the window pointer is being cleaned up at any rate.

Layers will call OnAttach instead of making the apps manually call it after pushing a new layer/overlay.