Update: ImGui
Added global define for using ImGui. Enabled for debug, release and distribution so far!

Think the crash for the build machine was happening when ImGui layer was cleaned up. There was another loop in the run on going and m_pImGuiLayer was nullptr. Added check around that before trying to call Begin/End!

Checking if GB_IMGUI_ENABLED is on for any ImGui related code in GB.