Update: Objects, Components + Graphics Contexts
Added base component class and empty sprite component class.
Objects have Initialize, Begin, Update, End and Deinitialize functions to use/override.
Renderer setup and graphics context class implemented.
Currently only supporting BGFX graphics context, this is just a wrapper essentially if I did ever want to try and implement my own renderer setup!
Added new macros for checking if pointers are valid or not:
 - GB_CHECK_PTR: Checks if a pointer is null, if so this will assert
 - GB_CHECK_PTR_SAFE: Checks if a pointer is null, if so, will just call return

The idea is to help write less lines of code with these macros but also help enforce correct pointer usage across the codebase. If a pointer is expected to always be valid during the time a function is called, using GB_CHECK_PTR would be the macro to use. However, if it's valid for the pointer to be null and you don't want to make the application crash, then using GB_CHECK_PTR_SAFE is the macro to use in these cases.

Prep work for implementing sprites.