Update: Editor Level Objects
Drawing a list of objects from the current editor level in the Level Hierarchy window (just displays whatever the name of the object is). The details window will also just display whatever the name of the selected object is. Can right click and select "Delete Object" to remove objects from the level or you can press the Delete key on the selected object to remove it!