Update: Commit Checker
:construction: :hammer:
Adding the CommitChecker tool to this repo now that it's in a more finished state (may still go back to it and add more features/polish the UI some more).

Removing the Linux/Mac header tool binaries so that I can add them in the next commit along. They weren't being picked up as being changed/updated. So readding them in the next commit should do the trick.