Update: WIP spdlog + Reflection
Updated spdlog to v2.x branch (need to include cpp files to build source code).
Copied latest GBHeaderTool files across, GBBool implementation to be used to draw vector elements of bools to work around how the std::vector<bool>.
Simplied DrawMember functions in visualiser class so the reflected member version is a just an overload of the type descriptor version (with a custom name too which will be the name of the variable or the index of the element in the vector).